pyproject.toml
[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"


[project]
# PEP 621 project metadata
# See https://www.python.org/dev/peps/pep-0621/
authors = [
    {name = "Lennon Cook", email = "lvc@codetripping.net"},
]
version = "0.1.0"
dynamic = ["classifiers"]
requires-python = ">=3.9"
dependencies = [
    
    "advent-of-code-data",
    "python-dotenv",
    "parse",
    "more-itertools",
    "regex",
    "numpy",
    "dataslots",
    "invoke<2.0.0,>=1.4.1",
]
dev-dependencies = []
includes = []
name = "aoc-2020"
description = "Advent of Code 2020"

[project.urls]

[project.optional-dependencies]

[tool]
[tool.pdm]
